[Enhancement] Decouple all global variables from extern in true singleton (#325)
All the global variables * g_* were decoupled to true singletons, reducing the dependencies between the objects and simplifying the amount of calls and instances. They used to be partial singletons mixed with global variables, that way we have one, and only one instance that is accessed via the getInstance methods, that are now the * g_* functions.

Changes:

g_actions
g_chat
g_config
g_creatureEvents
g_database
g_databaseTasks
g_decay
g_dispatcher
g_events
g_game
g_globalEvents
g_luaEnvironment
g_modules
g_monsters
g_moveEvents
g_npc
g_npcs
g_RSA
g_scheduler
g_scripts
g_spells
g_talkActions
g_vocations
g_weapons

The * g_luaEnvironment has been decoupled using inline, new feature of c++17

+++ src/game/game.h
@@ -340,7 +340,7 @@ class Game
 		void playerRequestAddVip(uint32_t playerId, const std::string& name);
 		void playerRequestRemoveVip(uint32_t playerId, uint32_t guid);
 		void playerRequestEditVip(uint32_t playerId, uint32_t guid, const std::string& description, uint32_t icon, bool notify);
-		void playerApplyImbuement(uint32_t playerId, uint32_t imbuementid, uint8_t slot, bool protectionCharm);
+		void playerApplyImbuement(uint32_t playerId, uint16_t imbuementid, uint8_t slot, bool protectionCharm);
 		void playerClearImbuement(uint32_t playerid, uint8_t slot);
 		void playerCloseImbuementWindow(uint32_t playerid);
 		void playerTurn(uint32_t playerId, Direction dir);
